feat: cleaned code, simplified logic and made fixes
