+++ .env.development.example
@@ -5,3 +5,7 @@ REACT_APP_KEY_ALCHEMY_API_KEY=""
 REACT_APP_KEY_INFURA_API_KEY=""
 PINATA_API_KEY=XXXXXXXXXXXXX
 PINATA_API_SECRET_KEY=XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
+
+REACT_APP_TENDERLY_PROJECT=dxdao-proposal-simulation
+REACT_APP_TENDERLY_USER=dxdao
+REACT_APP_TENDERLY_ACCESS_KEY=XXXXXXXXXXXXX
