+++ src/Components/ActionsBuilder/OptionsList/OptionsList.styled.tsx
@@ -1,5 +1,10 @@
+import { Box } from 'Components/Primitives/Layout';
 import { Button } from 'old-components/Guilds/common/Button';
 import styled from 'styled-components';
+
+export const AddOptionWrapper = styled(Box)`
+  padding: 1rem;
+`;
 
 export const SimulationButton = styled(Button)`
   padding: 0.6rem 1.25rem;
