+++ src/Components/ActionsBuilder/OptionsList/SimulationModal/SimulationModal.styled.tsx
