+++ src/Components/ActionsBuilder/OptionsList/types.ts
@@ -1,0 +1,17 @@
+import { Option } from '../types';
+
+export interface OptionsListProps {
+  isEditable: boolean;
+  options: Option[];
+  onChange: (options: Option[]) => void;
+  addOption: () => void;
+  editOption: (option: Option) => void;
+}
+
+export enum SimulationState {
+  none,
+  pending,
+  allPassed,
+  someFailed,
+  error,
+}
